backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
motor==3.3.2
pymongo==4.6.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
pydantic==2.6.1
python-multipart==0.0.9
redis[hiredis]==5.0.1
